feat: auditoría completa backend + renovación UI/UX premium
Backend:
- 22+ endpoints nuevos (PATCH users, system admin, forgot-password,
  presigned-url, resolve, alertas por tipo, métricas, seller endpoints)
- Connection pooling en 17 servicios (pool_size=5, max_overflow=10)
- Keystore de producción para APK release

Mobile - UI/UX:
- Sistema de paletas dinámico (5 paletas: Ocean, Indigo, Rose, Emerald, Sunset)
- Login, Registro, Forgot Password rediseñados con theme-aware
- Alertas premium animadas (AppAlert) reemplazan Alert.alert en 217 llamadas
- Fix contraste de colores en 15+ pantallas
- Botón ver/ocultar contraseña

Mobile - Funcionalidad:
- Mapas con OpenStreetMap via WebMapView (sin API key)
- Keyboard handling en 16 pantallas (KeyboardScreen component)
- 6 pantallas conectadas a API real (antes usaban MOCK)
- 4 features implementados (calendarios, analytics, edit clínica)
- 7 servicios mobile nuevos (insurance, training, venues, grooming,
  funerary, sponsors, rides) + 14 pantallas
- Navegación actualizada en menú
- Caché y timeout en apiFetch
- react-query optimizado (staleTime 60s, gcTime 5min)

Mobile - Limpieza:
- Eliminados archivos debug (test-export.js, ts_errors.txt, SafeMapView)
- Mocks sin uso removidos de directorio.ts
- Rutas de import corregidas (@/src/components/AppAlert)
- APK excluido de git (>100MB)

+++ backend/michicondrias_perdidas/app/models/petfriendly.py
@@ -22,7 +22,7 @@ class PetfriendlyPlace(Base):
 
     phone = Column(String(30), nullable=True)
     website = Column(String(255), nullable=True)
-    rating = Column(Integer, default=0)  # 1-5
+    rating = Column(Float, default=0)  # 1-5
 
     pet_sizes_allowed = Column(String(60), default="todos")  # pequeños, medianos, grandes, todos
     has_water_bowls = Column(String(10), default="no")
